+++ CHANGES.rst
@@ -5,6 +5,8 @@ CHANGELOG for LaunchKey Python SDK
 -----
 
 * Updated CLI to support separate encryption and signature keys
+* Altered JOSE transport to ensure only the designated signing key is used for signing requests
+* Added the `add_encryption_private_key` method to all factories and deprecated `add_additional_private_key`
 
 3.8.1
 -----